commit the test app
in an integration test suite, it makes sense to generated an app with a
generator-koa version and commit it here to be tested against, knowing
that the module-backing for koa will be the cwd. publishing /
submoduling the generator-koa and mixing with `yo` package sounds
complex and doesn't seem like the right solution for now.